EXTENDED: Start a new game with only 31 crew.
The logic behind this is that the Vindicator starts its journey with a full compliment of crew, stumbles upon the Tobermoon and crews it. Then, while traversing HyperSpace towards Sol, the Vindicator comes under attack by a Slylandro Probe and the Tobermoon is hit, killing Captain Burton. Filling uo the Tobermoon is 18 crew + replacing Cap. Burton is 19 crew transferred from the Vindicator to the Tobermoon. 50 - 19 = 31