Modify some functions and add a few more to help make things easier
Added a range randomizer, a short function to find out if an alien race is dead, a function to pick the planet closest to a habitable zone and nudge it in place if the one choses is out of the habitable bounds, and a "Planet Byte" generator based on the current number of planets.
Modified FillOrbits() and matchWorld() to make them use NUmPlanets and Planet/MoonByte using macros when applicable.